#include <16.h>.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,14 +8,14 @@
 int main(void)
 {
 	int i = 0;
-	unsigned long int a = 0, b = 1, next = 0;
+	unsigned long long int a = 0, b = 1, next = 0;
 
 	while (i < 98)
 	{
 		next = a + b;
 		a = b;
 		b = next;
-		printf("%lu", next);
+		printf("%llu", next);
 
 		if (i < 97)
 			printf(", ");
